(I7) the ledger, and the sixteenth and seventeenth chr(92) catches
The wave's ledger, its eleven decisions and its sixteen carried remainders are in
`docs/memos/island-progress.md` under *Done — wave I7*, with a compact block in
`docs/ROADMAP.md`.

**Two chr(92) catches, and they are different things wearing the same shape.**

* **A COLLISION.** `IslandReport`'s summary is a fixed-column table an author
  reads, and its alignment is runs of spaces INSIDE string literals — exactly the
  shape the sweep looks for. The sanctioned remedy is `ALIGNED_ON_PURPOSE`, which
  is keyed on the **enclosing function's name alone** — and there are **twenty-
  four** functions called `summary` in this workspace, so exempting the natural
  name would have exempted all of them. The table moved into a uniquely named
  private helper instead. An over-broad exemption is the ban-list hazard turned
  around.
* **A DEFECT, and the P22 law met head-on.** A scripted edit written through a
  Python heredoc put a `\`-continuation inside an `assert!` message; a lone
  backslash before a newline in a non-raw Python string is a **Python**
  continuation, so it ate the backslash and left eighteen spaces inside the
  literal. Caught by the sweep in the final battery, repaired through the Edit
  tool. The law is three waves old and has caught something in every one.

Fourteen clippy findings cleared, all in this wave's own code, and one rustdoc
warning (a public constant linking a private `smooth`) — so the wave adds **zero**
of either.

**Closing counts, measured at this head:** battery **318 / 5 946 / 0 / 16**
(six new test binaries, **73** new arms, two new `#[ignore]`s that need a fetched
cache CI does not have); frontend **702 / 78** with `tsc` and `eslint` clean;
goldens **54, byte-identical under `INF_GOLDEN_STRICT=1`** over 101 arms with no
PNG rewritten; `clippy --workspace --all-targets` **0**; rustdoc **404 over 46
documented crates**, measured CI-style after `cargo clean --doc` — the same 404
the I6 audit closed on, over one more crate; `cargo fmt --all --check` clean;
**no schema moved**; committed levels **21 → 23**.

`target/` reached **289 GB** mid-wave and the volume hit 100 %. `cargo clean`
freed **299.7 GiB** across 48 632 files — the phase-boundary clean the disk law
prescribes, taken at a phase boundary.

+++ crates/inf-island/src/source.rs
@@ -598,7 +598,7 @@ mod tests {
         assert_eq!(m.sea_level_tiles().len(), 1);
         assert!(m.sea_level_tiles().contains(&(1, 1)));
         // The fraction counts SAMPLES, which is what an extent-level policy reads.
-        let px = tilemath::TILE_PX as usize;
+        let px = tilemath::TILE_PX;
         assert!(
             (m.sea_level_fraction() - 0.5).abs() < 0.01,
             "one flat tile of two is about half the samples, got {}",
